clean pplcnet code
