+++ ppdet/modeling/reid/__init__.py
@@ -16,12 +16,10 @@ from . import jde_embedding_head
 from . import fairmot_embedding_head
 from . import resnet
 from . import pyramidal_embedding
-from . import pplcnet
 from . import pplcnet_embedding
 
 from .fairmot_embedding_head import *
 from .jde_embedding_head import *
 from .resnet import *
 from .pyramidal_embedding import *
-from .pplcnet import *
 from .pplcnet_embedding import *
